Fix bare excepts reported by pylint.
A bare except also catches exceptions like SystemExit and KeyboardInterrupt
that should only handled in the outer most exception handler if at all.
Otherwise Ctrl-C might not work to exit a Python program as expected if it
happened to be pressed while in a bare exception handler. In most places,
I changed it to catch Exception which is pretty general, but seems to be ok
to me in those cases where an exception should never lead to Nengo exiting.
Within the cache I added logging statements to provide a bit more debugging
information because those exception can indicate problems, but are not
a failure we cannot recover from (and we should recover from it).

I explicitly disabled the broad-except warning because we have quite a number
where it seems acceptable to catch broadly. I hope that disallowing broad
excepts and thus forcing to specify exception types, everyone will think about
whether `Exception` is appropriate or a more narrow scope would be better.